refactor: decompose lexer into functional submodules
Break down the lexer implementation into logical components to improve
code organization and readability.

Changes:
- **New Module Structure**:
  - `core.rs`: `Lexer` and `Token` struct definitions and entry points.
  - `cursor.rs`: Low-level source navigation (`advance`, `peek`, `match_next`).
  - `scan.rs`: Main token dispatch logic (`next_token`).
  - `ident.rs`: Identifier scanning and keyword mapping.
  - `literals.rs`: String and character literal parsing.
  - `trivia.rs`: Whitespace and comment skipping.
  - `common.rs`: Internal shared imports.
- **Integration**:
  - Updated `front/lexer/src/lib.rs` and `mod.rs` to expose the new structure.
  - Updated imports in `front/parser` to align with the refactored lexer API (explicit `use lexer::token::TokenType` where necessary).

This modularization separates concerns, making the lexer easier to maintain
and extend.
