+++ front/lexer/src/lexer/core.rs
@@ -1,0 +1,51 @@
+use crate::token::TokenType;
+use super::common::*;
+
+#[derive(Debug, Clone)]
+pub struct Token {
+    pub token_type: TokenType,
+    pub lexeme: String,
+    pub line: usize,
+}
+
+impl Token {
+    pub fn new(token_type: TokenType, lexeme: String, line: usize) -> Self {
+        Token { token_type, lexeme, line }
+    }
+}
+
+impl Default for Token {
+    fn default() -> Self {
+        Token {
+            token_type: TokenType::Eof,
+            lexeme: String::new(),
+            line: 0,
+        }
+    }
+}
+
+#[derive(Debug)]
+pub struct Lexer<'a> {
+    pub source: &'a str,
+    pub current: usize,
+    pub line: usize,
+}
+
+impl<'a> Lexer<'a> {
+    pub fn new(source: &'a str) -> Lexer<'a> {
+        Lexer { source, current: 0, line: 1 }
+    }
+
+    pub fn tokenize(&mut self) -> Vec<Token> {
+        let mut tokens = Vec::new();
+        loop {
+            let token = self.next_token(); // scan.rs에 구현
+            if token.token_type == TokenType::Eof {
+                tokens.push(token);
+                break;
+            }
+            tokens.push(token);
+        }
+        tokens
+    }
+}
