+++ front/lexer/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod lexer;
+pub mod token;
 
 pub use lexer::*;
