+++ front/parser/src/parser/control.rs
@@ -1,6 +1,7 @@
 use std::iter::Peekable;
 use std::slice::Iter;
-use lexer::{Token, TokenType};
+use lexer::Token;
+use lexer::token::TokenType;
 use crate::ast::{ASTNode, Expression, StatementNode};
 use crate::format::parse_expression;
 use crate::parser::stmt::parse_block;
