+++ front/parser/src/parser/expr.rs
@@ -1,6 +1,7 @@
 use std::iter::Peekable;
 use std::slice::Iter;
-use lexer::{Token, TokenType};
+use lexer::Token;
+use lexer::token::TokenType;
 use crate::ast::Expression;
 use crate::format::parse_expression;
