+++ front/parser/src/type_system.rs
@@ -1,6 +1,7 @@
 use crate::ast::WaveType;
 use crate::*;
-use lexer::{Token, TokenType};
+use lexer::Token;
+use lexer::token::TokenType;
 use std::iter::Peekable;
 use std::slice::Iter;
